new component user main view, refactor app
