+++ client/src/App.js
@@ -1,43 +1,47 @@
-import { Routes, Route } from 'react-router-dom'
 import { useEffect } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { setUser } from './reducers/loginFormReducer'
+import {
+  useNavigate,
+  Link,
+  Routes,
+  Route,
+  Navigate
+} from 'react-router-dom'
 import AdminProfile from './components/AdminProfile'
 import UserProfile from './components/UserProfile'
 import RegisterForm from './components/RegisterForm'
 import Notification from './components/Notification'
 import LoginForm from './components/LoginForm'
 import UserList from './components/UserList'
-import Home from './components/Home'
 import userService from './services/userService'
 import jwt_decode from 'jwt-decode'
-import { Link } from 'react-router-dom'
+import UserMainView from './components/UserMainView'
+
 
 function App() {
   const user = useSelector((state) => state.loginForm.user)
-  console.log('user appissa',user)
   const dispatch = useDispatch()
-
   const isJsonWebTokenExpired = jwt => {
     const decodedToken = jwt_decode(jwt)
     const expiresAtMillis = decodedToken.exp * 1000
     return expiresAtMillis < Date.now()
   }
-  /*   useEffect(() => {
-    if (user) {
-      dispatch(setUser(user))
-    }
-  }, [user]) */
+
+  const navigate = useNavigate()
 
   useEffect(() => {
-    console.log('useEffect')
     const loggedUserJSON = window.localStorage.getItem('loggedUser')
     if (loggedUserJSON) {
       const parsedUser = JSON.parse(loggedUserJSON)
       if (!isJsonWebTokenExpired(parsedUser.token)) {
-        console.log(' usertoken Appissa', parsedUser.token)
         dispatch(setUser(parsedUser))
         userService.setToken(parsedUser.token)
+      }
+      if (parsedUser.role === 'admin') {
+        navigate('/admin')
+      } else {
+        navigate('/user')
       }
     }
   }, [])
@@ -55,23 +59,28 @@ function App() {
             onClick={() => userService.logoutLocalUser(dispatch)}
             data-cy='logout'
           >
-            Kirjaudu ulos
+            Kirjaudu ulos käyttäjältä {user.firstName} {user.lastName}
           </button>
           <Link style={padding} to="/">Etusivu</Link>
-          { (user.role === 'admin') &&
-          <>
-            <Link style={padding} to="/users">Käyttäjät</Link>
-            <Link style={padding} to="/register">Luo käyttäjä</Link>
-          </>
+          { user.role === 'admin'
+            ?
+            <>
+              <Link style={padding} to="/users">Käyttäjät</Link>
+              <Link style={padding} to="/register">Luo käyttäjä</Link>
+            </>
+            :
+            <Link style={padding} to="/userprofile">Käyttäjätiedot</Link>
           }
           <Routes>
 
             <Route path="/admin" element={<AdminProfile />} />
-            <Route path="/user" element={<UserProfile />} />
-            <Route path="/" element={<Home />} />
+            <Route path="/user" element={<UserMainView />} />
+            <Route path="/userprofile" element={<UserProfile />} />
             <Route path="/users" element={<UserList />} />
             <Route path="/register" element={<RegisterForm />} />
-
+            <Route path="/" element={user.role === 'admin'
+              ? <Navigate replace to="/admin" />
+              : <Navigate replace to="/user" />}/>
           </Routes>
         </>
         : <LoginForm />
