+++ client/src/components/AdminProfile.js
@@ -1,5 +1,4 @@
 import { useSelector } from 'react-redux'
-//import { Link } from 'react-router-dom'
 
 const AdminProfile = () => {
   /**
@@ -8,15 +7,11 @@ const AdminProfile = () => {
                                * @type {userObject}
                                */
   const user = useSelector((state) => state.loginForm.user)
-
-  /*   const padding = {
-    padding: 5
-  } */
-
+  const date = new Date(user.expirationDate).toISOString().replace('-', '/').split('T')[0].replace('-', '/')
   return (
     <>
       <h2>Käyttäjienhallinta</h2>
-      <>Admin {user.username} kirjautunut</>
+      <>ADMININ KÄYTTÄJÄTUNNUS VANHENEE:  { date }</>
     </>
   )
 }
